fix: minimize Codex auth shared changes
